Added Functions and procedures, in addition to reorgnizing the files
