+++ RestaurantReservation.Db/RestaurantRevenue.cs
@@ -1,0 +1,6 @@
+﻿namespace RestaurantReservation.Db;
+
+public class RestaurantRevenue
+{
+    public int? Revenue { get; set; }
+}
